+++ packages/api/src/business/auth/objects/LoginArgs.ts
@@ -1,11 +1,15 @@
-import { MinLength } from "class-validator";
+import { IsEmail, MinLength } from "class-validator";
 import { Field, InputType } from "type-graphql";
 
 @InputType()
 export class LoginArgs {
   @Field(() => String)
-  @MinLength(1)
-  username: string = "";
+  @IsEmail({
+    allow_display_name: false,
+    allow_ip_domain: true,
+    require_tld: true,
+  })
+  email: string = "";
 
   @Field(() => String)
   @MinLength(1)