+++ packages/client/src/ui/LoginPage/LoginPage.tsx
@@ -23,7 +23,7 @@ export const LoginPage: FC = () => {
   usePageTitle("CloverCoin Species - Login");
   const error = useRouteActionData<"root.login">();
 
-  const [username, setUsername] = useState("");
+  const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
 
   useEffect(() => {
@@ -32,7 +32,7 @@ export const LoginPage: FC = () => {
       username: "",
       form: "",
     }));
-  }, [username]);
+  }, [email]);
 
   useEffect(() => {
     setFieldErrors((fieldErrors) => ({
@@ -42,11 +42,11 @@ export const LoginPage: FC = () => {
     }));
   }, [password]);
 
-  type FieldNames = "username" | "password" | "form";
+  type FieldNames = "email" | "password" | "form";
   const [fieldErrors, setFieldErrors] = useState<
     Record<FieldNames, string | undefined | null>
   >(() => ({
-    username: null,
+    email: null,
     password: null,
     form: null,
   }));
@@ -103,15 +103,15 @@ export const LoginPage: FC = () => {
         )}
         <br />
         <TextField
-          name="username"
-          label="Username"
-          type="text"
+          name="email"
+          label="Email"
+          type="email"
           required
-          error={Boolean(fieldErrors.username)}
-          helperText={fieldErrors.username ?? ""}
+          error={Boolean(fieldErrors.email)}
+          helperText={fieldErrors.email ?? ""}
           fullWidth
-          value={username}
-          onChange={({ target: { value } }) => setUsername(value)}
+          value={email}
+          onChange={({ target: { value } }) => setEmail(value)}
           css={ss.field}
         />
         <TextField
@@ -133,7 +133,7 @@ export const LoginPage: FC = () => {
           variant="contained"
           color="primary"
           disabled={
-            !username ||
+            !email ||
             !password ||
             // if there are errors, disabled
             Object.values(fieldErrors).some((val) => val)